requirements.txt
customtkinter
openai
pandas
pdfplumber
docx2txt
gspread
google-auth
python-dotenv
tiktoken
pymupdf
pillow
